[IMP] NEW FIELD QTY TOTAL

+++ sine_motoscoot_stock/product_motoscoot.py
@@ -109,7 +109,7 @@ class product_product(osv.osv):
                                        digits_compute=dp.get_precision('Precio Base TT (Tarifa Fabricante sin IVA)')),
         'internet': fields.boolean('Internet?', help='Está activo en Magento?'),
         'label_print': fields.boolean('Label Print?', help='Se debe imprimir la etiqueta en albaranes de entrada?'),
-        'qty_total': fields.function(compute_total_qty, type='integer', string='Stock Total'),
+        'qty_total': fields.function(compute_total_qty, type='float', string='Stock Total'),
     }
 
     _defaults = {
